+++ src/unstyle.rs
@@ -50,3 +50,42 @@ impl Unstyle for AsciiEscapedString {
         )
     }
 }
+
+#[cfg(feature = "ansi_term_style")]
+impl Unstyle for ansi_term::ANSIStrings<'_> {
+    fn unstyle(&self) -> String {
+        ansi_term::unstyle(self)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn string() {
+        let s = String::from("string with \nnewlines and some \ttabs\t\n\t");
+        assert_eq!(s.to_string(), s.unstyle());
+        assert_eq!(3, s.nb_of_lines());
+    }
+
+    #[test]
+    fn ascii_escaped_string() {
+        let s =
+            AsciiEscapedString::from("This is \x1b[1;31;44mbold red with blue background\x1b[0m");
+        assert_eq!("This is bold red with blue background", s.unstyle());
+    }
+
+    #[cfg(feature = "ansi_term_style")]
+    #[test]
+    fn ansi_term_unstyle() {
+        use ansi_term::Colour::Red;
+        use ansi_term::{ANSIString, ANSIStrings};
+
+        let some_value = format!("{:b}", 42);
+        let strings: &[ANSIString<'static>] =
+            &[Red.paint("["), Red.bold().paint(some_value), Red.paint("]")];
+
+        assert_eq!("[101010]", ANSIStrings(&strings).unstyle());
+    }
+}